--- insight_story/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'insight_story'
