--- insight_story/src/main/java/com/insight/pak/InsightStoryApplication.java
@@ -1,13 +1,0 @@
-package com.insight.pak;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class InsightStoryApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(InsightStoryApplication.class, args);
-	}
-
-}
