--- insight_story/src/main/resources/application.properties
@@ -1,8 +1,0 @@
-spring.application.name=insight_story
-
-# thymeleaf 참조 경로
-spring.thymeleaf.prefix=classpath:/templates/
-spring.thymeleaf.suffix=.html
-
-# templates 디렉토리에 파일이 있는지 없는지 체크, 없으면 에러를 발생시킨다.
-spring.thymeleaf.check-template-location=true
