--- insight_story/src/main/resources/templates/index.html
@@ -1,194 +1,0 @@
-<!doctype html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <title>Test Page</title>
-	<link rel="stylesheet" href="/index.css">
-
-	<script>
-	    function getUrlParameter(name) {
-	        name = name.replace(/[\[]/, '\\[').replace(/[\]]/, '\\]');
-	        var regex = new RegExp('[\\?&]' + name + '=([^&#]*)');
-	        var results = regex.exec(location.search);
-	        return results === null ? '' : decodeURIComponent(results[1].replace(/\+/g, ' '));
-	    };
-		
-		function goBack() {
-		        window.history.back();
-		};
-
-	    window.onload = function() {
-	        var keyword = getUrlParameter('keyword');
-	        if (keyword) {
-	            var keywordMap = {
-	                'martial-arts': '천하를 꿈꾸는 자',
-	                'fantasy': '왕녀의 복수',
-	                'mystery': '베이커가의 비밀',
-	                'horror': '어둠 속의 속삭임',
-	                'romance': '다시 만난 우리',
-	                'sf': '꿈의 경계'
-	            };
-				
-				var storyMap = {
-				                    'martial-arts': `무림의 세계에서 부모와 집, 모든 재산을 산적에게 잃고
-									떠돌아다니던 진산월. 그는 무당파에 들어가기 위해 무당파 본산 대문 앞에서
-									이틀째 기다리고 있다. 재치 있고 호기심 많은 성격의 산월은 잃어버린
-									모든 것을 되찾기 위해 새로운 길을 찾고자 한다.
-									
-									어느 날, 무당파의 문이 열리고 산월은 그의 운명을 바꿀 기회를 맞이하게 된다.
-									무림의 세계에 발을 들인 그는 다양한 인물들과 만나고, 새로운 기술을 익히며,
-									잃어버린 가족의 복수를 위해 힘을 키워나간다.
-									
-									하지만 무당파에서도 그의 도전은 끝나지 않는다. 산월은 무림의
-									어두운 비밀을 밝혀내고, 자신만의 방식으로 정의를 구현하려 한다.
-									그 과정에서 그는 친구와 적을 구분하며, 무림의 질서를
-									바로잡기 위해 고군분투한다.
-									
-									재치와 용기로 무림을 헤쳐나가는 진산월의 이야기
-									 "천하를 꿈꾸는 자"는 무협의 세계에서 펼쳐지는 그의 모험과 성장,
-									 그리고 정의를 위한 끝없는 도전을 그린다.`,
-									 
-				                    'fantasy': `중세 시대의 어느 왕국, 전대 왕비의 딸로 왕녀였던 브라이트 아실리는
-									어머니가 누명을 쓰고 처형당한 후, 거지촌에서 비참한 삶을 살게 된다.
-									자신이 쫓겨난 이유도 모른 채 살아가던 아실리는 우연히 의붓동생
-									다리아 블론디와 전대 용사 패트릭 안토니를 만나게 된다.
-
-									 현대 왕녀인 블론디는 어머니인 현대 왕비가 전대 왕비에게 죄를
-									 뒤집어씌운 사실을 알고 있다. 아실리를 거지촌에서 우연히 마주쳤지만,
-									 모른척하며 진실을 숨긴다.
-
-									 한편, 용을 쓰러트린 전대 용사 안토니는 아실리를 사랑했으나,
-									 그녀가 쫓겨난 이유로 함께할 수 없었다. 그는 현대 왕비에게
-									 복수할 날을 기다리며 블론디의 곁에서 기회를 엿본다.
-
-									 브라이트 아실리는 자신의 비참한 운명을 극복하고 진실을 밝혀내기 위해,
-									 안토니와 함께 현대 왕비에게 맞서 싸우기로 결심한다. 이들은 복수를
-									 통해 정의를 되찾고, 왕국의 진정한 평화를 이루기 위해 힘을 합치는데...`,
-									
-				                    'mystery': `19세기 런던, 명석한 두뇌와 뛰어난 관찰력으로
-									유명한 탐정 셜록 홈즈는 베이커가 221B에 거주하며
-									각종 사건을 해결해 명성을 쌓아가고 있다. 그의 절친한 친구이자
-									동료인 존 왓슨은 베테랑 군의관 출신으로, 홈즈의 사건 해결을 돕고
-									그 과정을 기록한다. 의학 지식을 활용해 홈즈의 탐정 업무에 큰 도움을
-									주는 왓슨은 그의 신뢰받는 파트너이다.
-
-									런던을 배경으로 펼쳐지는 미스터리한 사건들 속에서,
-									홈즈와 왓슨은 서로의 지식을 결합해 수많은 난제를 해결해 나간다.
-									살인, 실종, 도난 등 다양한 사건들이 두 사람 앞에 펼쳐지며,
-									그들의 우정과 협력은 더욱 깊어진다.
-
-									셜록 홈즈는 뛰어난 추리력으로 사건의 본질을 꿰뚫어보고,
-									왓슨은 그의 곁에서 충실히 지원한다. 그들은 함께 런던의
-									어두운 골목과 고풍스러운 저택을 탐험하며 진실을 찾아 나선다.
-									"베이커가의 비밀"은 뛰어난 탐정과 그의 파트너가 런던을 무대로
-									스릴 넘치는 추리와 우정의 이야기를 그린다.`,
-									
-				                    'horror':`1970년대 미국의 작은 마을, 제인 도는 부모가 살해당한 후
-									특별한 능력으로 귀신과 소통하며 할머니와 함께 살아간다.
-									 제인은 할머니의 요청으로 유명한 퇴마사
-									 마이클 스미스에게 퇴마 기술을 배우기 시작한다.
-									 
-									한편, 가족을 잃은 후 귀신과 싸워온 마이클은 제인의 스승이 되어
-									그녀에게 자신의 지식을 전수한다. 제인의 가장 친한 친구인 에밀리 존슨은
-									심령현상에 깊은 관심을 가지며, 제인을 지지하고 함께 초자연적인 현상을 조사한다.
-									 
-									이들 세 사람은 마을에서 일어나는 기이한 사건들을 해결해 나가며,
-									어둠의 세력과 맞서 싸운다. 수많은 도전과 위험 속에서,
-									이들은 서로의 유대감을 강화하며 마을을 지키기 위해 고군분투한다.
-									
-									귀신과의 싸움, 어둠의 힘에 맞서는 퇴마사들의 이야기.
-									1970년대 미국의 어두운 마을에서 펼쳐지는 공포와 스릴러가 가득한 이야기,
-									"어둠 속의 속삭임"이 지금 시작된다.`,
-									
-				                    'romance': `현대 대학교를 배경으로, 성실하고 조용한 성격의 서윤지는
-									유학 후 돌아와 다시 학교 생활을 시작한다. 대학 3학년인 윤지는
-									예전과 달라진 주변 사람들과의 복잡한 관계에 얽히게 된다.
-
-									강준호는 차가운 성격과 뛰어난 두뇌를 가진 대학생으로,
-									겉보기에는 완벽해 보이지만 어두운 면모를 숨기고 있다.
-									윤지와 가까워지며 그녀에게 특별한 관심을 보이는 준호는
-									자신만의 비밀을 간직한 채 그녀의 삶에 깊숙이 들어오게 된다.
-
-									윤지의 고등학교 동창인 이태민은 활발하고 사교적인 성격을 가지고 있다.
-									오랜만에 윤지를 만나게 된 태민은 그녀를 돕고 보호하려 하며,
-									윤지와의 우정을 다시 한 번 확인하게 된다.
-
-									복잡하게 얽힌 인연 속에서, 이들은 각자의 상처와
-									비밀을 마주하며 사랑과 우정을 키워나간다.
-									현대 대학교에서 펼쳐지는 로맨스와 드라마가 가득한 이야기,
-									"다시 만난 우리"가 지금 시작된다.`,
-									
-				                    'sf': `근미래의 도시, 사람들의 꿈은 더 이상 단순한 무의식의 영역이 아니다.
-									뛰어난 기술력을 가진 기업가 데이비드 크레인은 사람들의 꿈에 들어가
-									정보를 훔칠 수 있는 능력을 가지고 있다. 하지만 아내를 잃은 후
-									죄책감에 시달리며, 그는 자신의 능력을 더 이상 사용하지 않기로 결심한다.
-
-									한편, 심리학자 엘리자베스 로렌스는 사람들의 꿈을 분석하는
-									특별한 능력을 지녔다. 그녀는 데이비드에게 새로운 임무를 제안하며
-									그의 능력을 다시 사용하게 하려 한다. 그 임무는 바로, 꿈속에 숨어 있는 위험을
-									제거하고 사람들의 무의식을 지키는 것이다.
-
-									전직 군인 마크 스티븐스는 데이비드의 팀에 합류해 꿈속 임무를 수행한다.
-									그의 빠른 판단력과 행동력은 팀의 중요한 자산이다.
-
-									이제, 데이비드, 엘리자베스, 마크는 꿈의 경계를 넘나들며
-									사람들의 무의식을 지켜내야 한다. 꿈과 현실의 경계가 흐려지는 세계에서,
-									그들은 어떠한 선택을 하게 될 것인가? 한편, 꿈속에서 마주하는 위험은
-									그들 각자의 트라우마와 맞물려 예기치 못한 방향으로 흘러가는데...`
-
-				};
-				
-				var imageMap = {
-					'martial-arts': '1.jpg',
-					'fantasy': '2.jpg',
-					'mystery': '3.jpg',
-					'horror': '4.jpg',
-					'romance': '5.jpg',
-					'sf': '6.jpg'
-				};
-								
-				document.getElementById('keywordDisplay').innerText = keywordMap[keyword] || '';
-				document.getElementById('storyDisplay').innerText = storyMap[keyword] || '';
-				
-				var image_box = document.getElementById('image_box');
-				                if (imageMap[keyword]) {
-				                    image_box.style.backgroundImage = 'url(' + imageMap[keyword] + ')';
-				                } else {
-				                    image_box.style.backgroundImage = '';
-				                }
-	        }
-	    };
-	</script>
-  </head>
-  <body>
-	<div class="navbar">
-	    <div class="navbar-logo">
-	        <img src="insight-logo.png" alt="로고">
-	        <p>Insight story</p>
-	    </div>
-	</div>
-	
-	<div id="contents"> 
-		<div class="container">
-			<div class="box">
-				<div class="img_box" id="image_box">
-				</div>
-				<div class="span_box">
-					<span class="close" onclick="goBack()" style="cursor:pointer;">&times;</span>
-				</div>
-				<div class="keyword_box">
-					<h1 id="keywordDisplay"></h1>
-				</div>
-				<div class="story_box">
-						<p id="storyDisplay"></p>
-				</div>
-				<div class="start_box">
-					<div class="start" onclick="window.location.href='/content'" style="cursor:pointer;">
-						시작하기
-					</div>
-				</div>
-			</div>
-		</div>
-	</div>
-	</body>
-</html>
