--- insight_story/src/test/java/com/insight/pak/InsightStoryApplicationTests.java
@@ -1,13 +1,0 @@
-package com.insight.pak;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class InsightStoryApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
