+++ springboot/insight_story/src/main/resources/static/index.css
