+++ springboot/insight_story/src/main/resources/static/root_page.css
