+++ springboot/insight_story/src/main/resources/templates/content.html
