+++ springboot/insight_story/src/main/resources/templates/root_page.html
@@ -3,40 +3,86 @@
   <head>
     <meta charset="utf-8" />
     <title>Root Page</title>
-	<link rel="stylesheet" href="/css/root_page.css">
-  </head>
+	<link rel="stylesheet" href="/root_page.css">
+	<!-- alert 창 꾸미기 -->
+	<script src="https://unpkg.com/sweetalert/dist/sweetalert.min.js"></script> 
+	<script>
+	    // 모달 열기
+	    window.onload = function() {
+	        var modal = document.getElementById("keyModal");
+	        modal.style.display = "block";
+
+	        // 모달 창이 로드되었을 때 메시지 수신
+	        window.addEventListener('message', function(event) {
+				if (event.data === 'closeModal') {
+					modal.style.display = "none";
+				} else if (event.data === 'error') {
+					swal({
+						title: "오류!",
+						text: `OpenAi API Key는 정확히 56자여야 합니다!
+						키를 다시 확인해 주세요.`,
+						icon: "warning",
+						button: "확인",
+						dangerMode: true
+					});
+				} else if (event.data === 'success') {
+					swal({
+					  title: "확인 완료!",
+					  text: "OpenAi API Key가 확인되었습니다!",
+					  icon: "success",
+					  button: "확인",
+					});
+				}
+	        });
+	    };
+		
+		function goToNextPage(keyword) {
+		           window.location.href = '/index?keyword=' + encodeURIComponent(keyword);
+		        }
+				
+	</script>
+</head>
   <body>
-    <div id="header">
-		<div class="logo">
-				<img src="/img/test_logo.png" title="테스트 로고"/>
-		</div>
+	<div class="navbar">
+	    <div class="navbar-logo">
+	        <img src="insight-logo.png" alt="로고">
+	        <p>Insight story</p>
+	        <form id="storyForm" th:action="@{/root_page}" method="post">
+	        </form>
+	    </div>
 	</div>
 	
 	<div id="contents"> 
 		<div class="container">
-			<div class="left_box" onclick="location.href=('https:www.naver.net')" style="cursor:pointer;">
-				#키워드1
+			<div class="left_box" onclick="goToNextPage('martial-arts')" style="cursor:pointer;">
+				#무협
 			</div>
-			<div class="right_box">
-				#키워드2
-			</div>
-		</div>
-		<div class="container">
-			<div class="left_box">
-				#키워드3
-			</div>
-			<div class="right_box">
-				#키워드4
+			<div class="right_box" onclick="goToNextPage('horror')" style="cursor:pointer;">
+				#공포
 			</div>
 		</div>
 		<div class="container">
-			<div class="left_box">
-				#키워드5
+			<div class="left_box" onclick="goToNextPage('fantasy')" style="cursor:pointer;">
+				#판타지
 			</div>
-			<div class="right_box">
-				#키워드6
+			<div class="right_box" onclick="goToNextPage('romance')" style="cursor:pointer;">
+				#로맨스
 			</div>
 		</div>
+		<div class="container">
+			<div class="left_box" onclick="goToNextPage('mystery')" style="cursor:pointer;">
+				#추리
+			</div>
+			<div class="right_box" onclick="goToNextPage('sf')" style="cursor:pointer;">
+				#SF
+			</div>
+		</div>
+	</div>	
+	
+	<div id="keyModal" class="modal">
+	    <div class="modal-content">
+	        <iframe src="api_key" frameborder="0" style="width: 100%; height: 232px;"></iframe>
+	    </div>
 	</div>
   </body>
 </html>
